backend/requirements.txt
Flask>=3.0
Flask-Cors>=4.0
SQLAlchemy>=2.0
python-dotenv>=1.0
Pillow>=10.0
requests>=2.28
